Refactor SegmentationClassMap to it's own file
Also move  color utils to label_source/utils
