+++ src/rastervision/data/label_source/__init__.py
@@ -6,5 +6,6 @@ from rastervision.data.label_source.chip_classification_geojson_source import *
 from rastervision.data.label_source.chip_classification_geojson_source_config import *
 from rastervision.data.label_source.object_detection_geojson_source import *
 from rastervision.data.label_source.object_detection_geojson_source_config import *
+from rastervision.data.label_source.segmentation_class_transformer import SegmentationClassTransformer
 from rastervision.data.label_source.semantic_segmentation_raster_source import *
 from rastervision.data.label_source.semantic_segmentation_raster_source_config import *
