+++ src/rastervision/data/label_store/semantic_segmentation_raster_store_config.py
@@ -7,13 +7,16 @@ from rastervision.data.label_store import (
 
 
 class SemanticSegmentationRasterStoreConfig(LabelStoreConfig):
-    def __init__(self, uri=None):
+    def __init__(self, uri=None, rgb=False):
         super().__init__(store_type=rv.SEMANTIC_SEGMENTATION_RASTER)
         self.uri = uri
+        self.rgb = rgb
 
     def to_proto(self):
         msg = super().to_proto()
-        msg.uri = self.uri
+        if self.uri:
+            msg.semantic_segmentation_raster_store.uri = self.uri
+        msg.semantic_segmentation_raster_store.rgb = self.rgb
         return msg
 
     def for_prediction(self, label_uri):
@@ -22,8 +25,11 @@ class SemanticSegmentationRasterStoreConfig(LabelStoreConfig):
                    .build()
 
     def create_store(self, task_config, crs_transformer, tmp_dir):
-        return SemanticSegmentationRasterStore(self.uri, crs_transformer,
-                                               task_config.class_map, tmp_dir)
+        class_map = None
+        if self.rgb:
+            class_map = task_config.class_map
+        return SemanticSegmentationRasterStore(
+            self.uri, crs_transformer, tmp_dir, class_map=class_map)
 
     def preprocess_command(self, command_type, experiment_config, context=[]):
         conf = self
@@ -64,18 +70,25 @@ class SemanticSegmentationRasterStoreConfigBuilder(LabelStoreConfigBuilder):
     def __init__(self, prev=None):
         config = {}
         if prev:
-            config = {'uri': prev.uri}
+            config = {'uri': prev.uri, 'rgb': prev.rgb}
 
         super().__init__(SemanticSegmentationRasterStoreConfig, config)
 
     def from_proto(self, msg):
-        b = SemanticSegmentationRasterStoreConfigBuilder()
-
-        return b \
-            .with_uri(msg.uri)
+        return self.with_uri(msg.semantic_segmentation_raster_store.uri) \
+                   .with_rgb(msg.semantic_segmentation_raster_store.rgb)
 
     def with_uri(self, uri):
         """Set URI for a GeoTIFF used to read/write predictions."""
         b = deepcopy(self)
         b.config['uri'] = uri
+        return b
+
+    def with_rgb(self, rgb):
+        """Set flag for writing RGB data using the class map.
+
+        Otherwise this method will write the class ID into a single band.
+        """
+        b = deepcopy(self)
+        b.config['rgb'] = rgb
         return b
